slight refactor to clear-up code. File re-org: all helper functions stored in hooks folder as seperate files. initial setup to for new issue.
