+++ components/sections/faq.tsx
@@ -1,8 +1,8 @@
 export default function FAQ() {
 
     return (
-      <>
-      <h1>This is The FAQ</h1>
-      </>
+      <section>
+        <h1>This is The FAQ</h1>
+      </section>
     )
   }
