+++ components/sharedSections/footer.tsx
@@ -9,30 +9,28 @@ import SocialMediaIcons from "../elements/socialMediaIcons";
 export default function Footer({showSignUp}: {showSignUp: boolean}) {
 
     return (
-      <>
-        <footer >
-          <p className='font-quote text-center text-lg m-5'>{webLinks.quote}</p>
-          <div className="bg-black pt-10">
-            <InstagramShowcase customUrl={webLinks.instagram} height="112px"/>
-            <div className="my-10 md:flex md:w-full">
-              <SocialMediaIcons />
-              <div className="flex place-content-between mx-10 md:mt-auto md:w-full">
-                <div>
-                  <ul className="text-white font-body text-sm capitalize">
-                    {navLinks.map((pageInfo, index) => { return(
-                      <li key={index}><Link href={pageInfo.path}>{pageInfo.pageTitle}</Link></li>
-                    )})}
-                  </ul>
-                </div>
-                <Link href="/" className="self-end md:self-center">
-                  <CompactLogo color='#ffffff' className='block md:hidden h-14'/>
-                  <WebsiteLogo color='#ffffff' className="hidden md:block h-8"/>
-                </Link>
+      <footer >
+        <p className='font-quote text-center text-lg m-5'>{webLinks.quote}</p>
+        <div className="bg-black pt-10">
+          <InstagramShowcase customUrl={webLinks.instagram} height="112px"/>
+          <div className="my-10 md:flex md:w-full">
+            <SocialMediaIcons />
+            <div className="flex place-content-between mx-10 md:mt-auto md:w-full">
+              <div>
+                <ul className="text-white font-body text-sm capitalize">
+                  {navLinks.map((pageInfo, index) => { return(
+                    <li key={index}><Link href={pageInfo.path}>{pageInfo.pageTitle}</Link></li>
+                  )})}
+                </ul>
               </div>
+              <Link href="/" className="self-end md:self-center">
+                <CompactLogo color='#ffffff' className='block md:hidden h-14'/>
+                <WebsiteLogo color='#ffffff' className="hidden md:block h-8"/>
+              </Link>
             </div>
-            <p className="font-body text-xs md:text-sm text-center text-white py-2">Designed and Build with ❤️ by <Link href="https://fuentesnina.github.io/my-portfolio-website/">Nina Fuentes LLC</Link></p>
           </div>
-        </footer>
-      </>
+          <p className="font-body text-xs md:text-sm text-center text-white py-2">Designed and Build with ❤️ by <Link href="https://fuentesnina.github.io/my-portfolio-website/">Nina Fuentes LLC</Link></p>
+        </div>
+      </footer>
     )
   }
